🐛 fix(fonts): serif and display fonts not rendering and causing build to fail
