+++ src/app/layout.tsx
@@ -1,22 +1,8 @@
 import Footer from '@/components/layout/footer';
 import Header from '@/components/layout/header';
 import type { Metadata } from 'next';
-import { Dancing_Script, Nunito_Sans, Rozha_One } from 'next/font/google';
+import { nunitoSans, rhozaOne, dancingScript } from '@/utils/fonts';
 import './globals.css';
-
-const nunitoSans = Nunito_Sans({
-  subsets: ['latin'],
-  variable: '--font-nunito-sans'
-});
-export const rhozaOne = Rozha_One({
-  subsets: ['latin'],
-  variable: '--font-rhoza-one',
-  weight: ['400']
-});
-export const dancingScript = Dancing_Script({
-  subsets: ['latin'],
-  variable: '--font-dancing-script'
-});
 
 export const metadata: Metadata = {
   title: 'Crema e Cioccolato - Barletta',
@@ -29,7 +15,10 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang='it' className={nunitoSans.className}>
+    <html
+      lang='it'
+      className={`${nunitoSans.variable} ${rhozaOne.variable} ${dancingScript.variable}`}
+    >
       <body className='bg-babyPowder text-smokyBrown-900'>
         <Header />
         <main>{children}</main>
